+++ Library/ViewModels/ProjectSummaryCarouselViewModelTests.swift
@@ -1,4 +1,5 @@
 import Foundation
+import KsApi
 import Library
 import ReactiveExtensions
 import ReactiveExtensions_TestHelpers
@@ -6,7 +7,8 @@ import ReactiveExtensions_TestHelpers
 final class ProjectSummaryCarouselViewModelViewModelTests: TestCase {
   private let vm: ProjectSummaryCarouselViewModelType = ProjectSummaryCarouselViewModel()
 
-  private let loadProjectSummaryItemsIntoDataSource = TestObserver<[ProjectSummaryItem], Never>()
+  private let loadProjectSummaryItemsIntoDataSource
+    = TestObserver<[ProjectSummaryEnvelope.ProjectSummaryItem], Never>()
 
   override func setUp() {
     super.setUp()
@@ -18,7 +20,20 @@ final class ProjectSummaryCarouselViewModelViewModelTests: TestCase {
   func testLoadProjectSummaryItemsIntoDataSource() {
     self.loadProjectSummaryItemsIntoDataSource.assertDidNotEmitValue()
 
-    let items: [ProjectSummaryItem] = [1, 2, 3]
+    let items = [
+      ProjectSummaryEnvelope.ProjectSummaryItem(
+        question: .whatIsTheProject,
+        response: "Test copy 1"
+      ),
+      ProjectSummaryEnvelope.ProjectSummaryItem(
+        question: .whatWillYouDoWithTheMoney,
+        response: "Test copy 2"
+      ),
+      ProjectSummaryEnvelope.ProjectSummaryItem(
+        question: .whoAreYou,
+        response: "Test copy 3"
+      )
+    ]
 
     self.vm.inputs.configure(with: items)
